+++ apps/backend/src/adapters/hands_sea/dto.rs
@@ -6,6 +6,6 @@ use crate::repos::hands::Card;
 #[derive(Debug, Clone)]
 pub struct HandCreate {
     pub round_id: i64,
-    pub player_seat: i16,
+    pub player_seat: u8,
     pub cards: Vec<Card>,
 }
