+++ apps/backend/src/adapters/rounds_sea/dto.rs
@@ -6,9 +6,9 @@ use crate::entities::game_rounds::CardTrump;
 #[derive(Debug, Clone)]
 pub struct RoundCreate {
     pub game_id: i64,
-    pub round_no: i16,
-    pub hand_size: i16,
-    pub dealer_pos: i16,
+    pub round_no: u8,
+    pub hand_size: u8,
+    pub dealer_pos: u8,
 }
 
 /// DTO for updating trump selection.
