+++ apps/backend/src/adapters/tricks_sea/dto.rs
@@ -6,7 +6,7 @@ use crate::entities::round_tricks::CardSuit;
 #[derive(Debug, Clone)]
 pub struct TrickCreate {
     pub round_id: i64,
-    pub trick_no: i16,
+    pub trick_no: u8,
     pub lead_suit: CardSuit,
-    pub winner_seat: i16,
+    pub winner_seat: u8,
 }
